Moved utility functions in SleepScreen to separate file.
