+++ src/_utilities/RegexFunctions.js
@@ -1,0 +1,8 @@
+export const validCharLength = value => value.length > 7;
+
+export const lowercaseTest = (value) => /[a-z]/.test(value);
+
+export const uppercaseTest = (value) => /[A-Z]/.test(value);
+
+export const specialsTest = (value) =>
+  /[`!@#$%^&*()_+\-=\[\]{};':"\\|,.<>\/?~]/.test(value);
