+++ src/container/SleepScreen/SleepScreen.js
@@ -9,6 +9,7 @@ import images from '../../../assets/images';
 import WeeklyTable from '../../components/WeeklyTable/WeeklyTable';
 import EditSleepGoal from '../../components/EditSleepGoal/EditSleepGoal';
 import { styleConstants } from "../../_constants/StyleConstants";
+import {minutesToHours, bedtimeString} from "../../_utilities/UtilityFunctions";
 
 const weeklyAveSleep = 450;
 // define a hardcoded bedtime as a Date object with current date, set time to 10:30 PM
@@ -19,45 +20,6 @@ const sleepGoal = 480;
 const dailySleep = 450;
 const successWeek = [false, true, false, true, false, true, false];
 
-// helper function to convert minutes to H:MM format
-const minutesToHours = minutes => {
-  // define two text variables to be concatenated into a time string
-  let hours,
-    minText = '';
-  // calculate hours portion and convert to string
-  hours = Math.floor(minutes / 60).toString();
-  // calculate minutes portion
-  let min = minutes % 60;
-  // minutes string needs to be in a 2-digit format
-  if (min < 10) {
-    minText = '0' + min.toString();
-  } else {
-    minText = min.toString();
-  }
-  // return concatenation of hours and min portions separated by a colon
-  return hours + ':' + minText;
-};
-
-// helper function to convert bedtime date object to HH:MM AM/PM string
-const bedtimeString = bedtimeDate => {
-  // define hours and minutes
-  let hours = bedtimeDate.getHours();
-  let minutes = bedtimeDate.getMinutes();
-  let minText = '';
-  let ampm = 'PM';
-  if (hours < 12) {
-    ampm = 'AM';
-  } else {
-    hours -= 12;
-  }
-  // minutes string needs to be in a 2-digit format
-  if (minutes < 10) {
-    minText = '0' + minutes.toString();
-  } else {
-    minText = minutes.toString();
-  }
-  return hours.toString() + ':' + minText + ' ' + ampm;
-};
 
 export default function SleepScreen({navigation, user}) {
   console.log('In Sleep Screen with navigation:', navigation);
